+++ src/components/Menu/src/components/BasicMenuItem.vue
@@ -1,0 +1,16 @@
+<script lang="ts" setup>
+  import { Menu } from 'ant-design-vue';
+  import { itemProps } from '../props';
+  import MenuItemContent from './MenuItemContent.vue';
+
+  const MenuItem = Menu.Item;
+  defineProps(itemProps);
+</script>
+
+<template>
+  <MenuItem :key="item.path">
+    <MenuItemContent v-bind="$props" :item="item" />
+  </MenuItem>
+</template>
+
+<style lang="less" scoped></style>