+++ src/components/Menu/src/types.ts
@@ -1,0 +1,11 @@
+export interface MenuState {
+  defaultSelectedKeys: string[];
+
+  inlineIndent?: number;
+
+  openKeys: string[];
+
+  selectedKeys: string[];
+
+  collapsedOpenKeys: string[];
+}