+++ src/hooks/web/usePage.ts
@@ -1,0 +1,30 @@
+import type { Router, RouteLocationRaw } from 'vue-router';
+
+import { PageEnum } from '/@/enums/pageEnum';
+import { isString } from '/@/utils/is';
+import { useRouter } from 'vue-router';
+
+export type RouteLocationRawEx = Omit<RouteLocationRaw, 'path'> & { path: PageEnum };
+
+const handleError = (e: Error) => console.log(e);
+
+export function useGo(_router?: Router) {
+  let router;
+  if (!_router) {
+    router = useRouter();
+    console.log(router);
+  }
+  const { push, replace } = _router || router;
+  function go(opt: PageEnum | RouteLocationRawEx | string = PageEnum.BASE_HOME, isReplace = false) {
+    if (!opt) {
+      return;
+    }
+    if (isString(opt)) {
+      isReplace ? replace(opt).catch(handleError) : push(opt).catch(handleError);
+    } else {
+      const o = opt as RouteLocationRaw;
+      isReplace ? replace(o).catch(handleError) : push(opt).catch(handleError);
+    }
+  }
+  return go;
+}