+++ src/layouts/default/header/index.vue
@@ -1,0 +1,10 @@
+<script lang="ts" setup>
+  import { unref, computed } from 'vue';
+  import { propTypes } from '/@/utils/propTypes';
+  import { Layout } from 'ant-design-vue';
+  import { AppLogo } from '/@/components/Application';
+</script>
+
+<template> </template>
+
+<style lang="" scoped></style>