+++ src/layouts/default/menu/index.vue
@@ -1,0 +1,14 @@
+<script lang="tsx">
+  import type { PropType, CSSProperties } from 'vue';
+
+  import { computed, defineComponent, unref, toRef } from 'vue';
+  // import { BasicMenu } 
+</script>
+
+<template>
+
+</template>
+
+<style lang="less" scoped>
+
+</style>