+++ src/main.ts
@@ -1,6 +1,7 @@
 import '/@/design/index.less';
 import { createApp } from 'vue';
 import App from './App.vue';
+import { setupStore } from './store';
 import 'virtual:windi.css';
 import 'virtual:svg-icons-register';
 
@@ -8,5 +9,12 @@ if (import.meta.env.DEV) {
   import('ant-design-vue/dist/antd.less');
 }
 
-const app = createApp(App);
-app.mount('#app');
+async function bootstrap() {
+  const app = createApp(App);
+
+  setupStore(app);
+
+  app.mount('#app');
+}
+
+bootstrap();