+++ src/router/routes/basic.ts
@@ -1,1 +1,9 @@
-import type { AppRouteRecordRaw } from '/@/router/types'
+import type { AppRouteRecordRaw } from '/@/router/types';
+import { t } from '/@/hooks/web/useI18n';
+import {
+  REDIRECT_NAME,
+  LAYOUT,
+  EXCEPTION_COMPONENT,
+  PAGE_NOT_FOUND_NAME,
+} from '/@/router/constant';
+