+++ src/router/routes/index.ts
@@ -1,2 +1,3 @@
-import type { AppRouteRecordRaw, AppRouteModule } from 'vue-router';
+import type { AppRouteRecordRaw, AppRouteModule } from '../types';
 
+import { PAGE_NOT,  } from '../constant';