+++ src/router/types.ts
@@ -1,6 +1,7 @@
 import type { RouteRecordRaw, RouteMeta } from 'vue-router';
 // import { RoleEnum } from '../enums/roleEnum';
 import { defineComponent } from 'vue';
+import { RoleEnum } from '../enums/roleEnum';
 
 export type Component<T = any> =
   | ReturnType<typeof defineComponent>
@@ -17,3 +18,30 @@ export interface AppRouteRecordRaw extends Omit<RouteRecordRaw, 'meta'> {
   props?: Recordable;
   fullPath?: string;
 }
+
+export interface MenuTag {
+  type?: 'primary' | 'error' | 'warn' | 'success';
+  content?: string;
+  dot?: boolean;
+}
+
+export interface Menu {
+  name: string;
+  icon?: string;
+  path: string;
+  paramsPath?: string;
+  disabled?: boolean;
+  children?: Menu[];
+  orderNo?: number;
+  roles?: RoleEnum;
+  meta?: Partial<RouteMeta>;
+  tag?: MenuTag;
+  hideMenu?: boolean;
+}
+
+export interface MenuModule {
+  orderNo?: number;
+  menu: Menu;
+}
+
+export type AppRouteModule = AppRouteRecordRaw;