+++ src/store/modules/app.ts
@@ -1,4 +1,11 @@
-import type { ProjectConfig } from '/#/config';
+import type {
+  ProjectConfig,
+  HeaderSetting,
+  MenuSetting,
+  TransitionSetting,
+  MultiTabsSetting,
+} from '/#/config';
+import type { BeforeMiniState } from '/#/store';
 
 import { defineStore } from 'pinia';
 import { store } from '..';
@@ -7,3 +14,89 @@ import { APP_DARK_MODE_KEY_, PROJ_CFG_KEY } from '/@/enums/cacheEnum';
 import { Persistent } from '/@/utils/cache/persistent';
 import { darkMode } from '/@/settings/designSetting';
 import { deepMerge } from '/@/utils';
+// import { resetRouter } from '/@/router';
+
+interface AppState {
+  darkMode?: ThemeEnum;
+  pageLoading: boolean;
+  projectConfig: ProjectConfig | null;
+  beforeMiniInfo: BeforeMiniState;
+}
+
+let timeId: TimeoutHandle;
+export const useAppStore = defineStore({
+  id: 'app',
+  state: (): AppState => ({
+    darkMode: undefined,
+    pageLoading: false,
+    projectConfig: Persistent.getLocal(PROJ_CFG_KEY),
+    beforeMiniInfo: {},
+  }),
+  getters: {
+    getPageLoading(): boolean {
+      return this.pageLoading;
+    },
+    getDarkMode(): 'light' | 'night' | string {
+      return this.darkMode || localStorage.getItem(APP_DARK_MODE_KEY_) || darkMode;
+    },
+    getBeforeMiniInfo(): BeforeMiniState {
+      return this.beforeMiniInfo;
+    },
+    getProjectConfig(): ProjectConfig {
+      return this.projectConfig || ({} as ProjectConfig);
+    },
+    getHeaderSetting(): HeaderSetting {
+      return this.getProjectConfig.headerSetting;
+    },
+    getMenuSetting(): MenuSetting {
+      return this.getProjectConfig.menuSetting;
+    },
+    getMultiTabSetting(): MultiTabsSetting {
+      return this.getProjectConfig.multiTabsSetting;
+    },
+    getTransitionSetting(): TransitionSetting {
+      return this.getProjectConfig.transitionSetting;
+    },
+  },
+  actions: {
+    setPageLoading(loading: boolean): void {
+      this.pageLoading = loading;
+    },
+
+    setDarkMode(mode: ThemeEnum): void {
+      this.darkMode = mode;
+      localStorage.setItem(APP_DARK_MODE_KEY_, mode);
+    },
+
+    setBeforeMiniInfo(state: BeforeMiniState): void {
+      this.beforeMiniInfo = state;
+    },
+
+    setProjectConfig(config: DeepPartial<ProjectConfig>): void {
+      this.projectConfig = deepMerge(this.projectConfig || {}, config);
+      Persistent.setLocal(PROJ_CFG_KEY, this.projectConfig);
+    },
+
+    async resetAllState() {
+      // resetRouter();
+      Persistent.clearAll();
+    },
+
+    async setPageLoadingAction(loading: boolean): Promise<void> {
+      if (loading) {
+        clearTimeout(timeId);
+
+        timeId = setTimeout(() => {
+          this.setPageLoading(loading);
+        }, 50);
+      } else {
+        this.setPageLoading(loading);
+        clearTimeout(timeId);
+      }
+    },
+  },
+});
+
+export function useAppStoreWithOut() {
+  return useAppStore(store);
+}