+++ src/utils/request/index.ts
@@ -2,20 +2,20 @@
 // The axios configuration can be changed according to the project, just change the file, other files can be left unchanged
 
 import type { AxiosResponse } from 'axios';
-import type { RequestOptions, Result } from '/#/axios';
+import type { OtherResult, RequestOptions, Result } from '/#/axios';
 import type { AxiosTransform, CreateAxiosOptions } from './axiosTransform';
 import { VAxios } from './Axios';
 import { checkStatus } from './checkStatus';
 import { useGlobSetting } from '/@/hooks/setting';
 import { useMessage } from '/@/hooks/web/useMessage';
-import { RequestEnum, ResultEnum, ContentTypeEnum } from '/@/enums/httpEnum';
+import { RequestEnum, ContentTypeEnum } from '/@/enums/httpEnum';
 import { isString } from '/@/utils/is';
 import { getToken } from '/@/utils/auth';
 import { setObjToUrlParams, deepMerge } from '/@/utils';
 import { useErrorLogStoreWithOut } from '/@/store/modules/errorLog';
 import { useI18n } from '/@/hooks/web/useI18n';
 import { joinTimestamp, formatRequestDate } from './helper';
-import { useUserStoreWithOut } from '/@/store/modules/user';
+// import { useUserStoreWithOut } from '/@/store/modules/user';
 
 const globSetting = useGlobSetting();
 const urlPrefix = globSetting.urlPrefix;
@@ -28,7 +28,7 @@ const transform: AxiosTransform = {
   /**
    * @description: 处理请求数据。如果数据不是预期格式，可直接抛出错误
    */
-  transformRequestHook: (res: AxiosResponse<Result>, options: RequestOptions) => {
+  transformRequestHook: (res: AxiosResponse<Result | OtherResult>, options: RequestOptions) => {
     const { t } = useI18n();
     const { isTransformResponse, isReturnNativeResponse } = options;
     // 是否返回原生响应头 比如：需要获取响应头时使用该属性
@@ -47,40 +47,51 @@ const transform: AxiosTransform = {
       // return '[HTTP] Request has no return value';
       throw new Error(t('sys.api.apiRequestFailed'));
     }
-    //  这里 code，result，message为 后台统一的字段，需要在 types.ts内修改为项目自己的接口返回格式
-    const { code, result, message } = data;
+    // 这里为后台统一的字段，需要在 types.ts内修改为项目自己的接口返回格式
+    if ((data as OtherResult).content) {
+      const { state, content, message, success } = data as OtherResult;
 
-    // 这里逻辑可以根据项目进行修改
-    const hasSuccess = data && Reflect.has(data, 'code') && code === ResultEnum.SUCCESS;
-    if (hasSuccess) {
-      return result;
+      // 这里逻辑可以根据项目进行修改
+      const hasSuccess = data && Reflect.has(data, 'success') && success && state === 0;
+      if (hasSuccess) {
+        return content;
+      } else {
+        createErrorModal({ title: t('sys.api.errorTip'), content: message });
+      }
+    } else {
+      const { code, mesg, data: result } = data as Result;
+      if (code === '0') {
+        return result;
+      } else {
+        createErrorModal({ title: t('sys.api.errorTip'), content: mesg });
+      }
     }
 
     // 在此处根据自己项目的实际情况对不同的code执行不同的操作
     // 如果不希望中断当前请求，请return数据，否则直接抛出异常即可
-    let timeoutMsg = '';
-    switch (code) {
-      case ResultEnum.TIMEOUT:
-        timeoutMsg = t('sys.api.timeoutMessage');
-        const userStore = useUserStoreWithOut();
-        userStore.setToken(undefined);
-        userStore.logout(true);
-        break;
-      default:
-        if (message) {
-          timeoutMsg = message;
-        }
-    }
+    // let timeoutMsg = '';
+    // switch (state) {
+    //   case ResultEnum.TIMEOUT:
+    //     timeoutMsg = t('sys.api.timeoutMessage');
+    //     const userStore = useUserStoreWithOut();
+    //     userStore.setToken(undefined);
+    //     userStore.logout(true);
+    //     break;
+    //   default:
+    //     if (message) {
+    //       timeoutMsg = message;
+    //     }
+    // }
 
     // errorMessageMode=‘modal’的时候会显示modal错误弹窗，而不是消息提示，用于一些比较重要的错误
     // errorMessageMode='none' 一般是调用时明确表示不希望自动弹出错误提示
-    if (options.errorMessageMode === 'modal') {
-      createErrorModal({ title: t('sys.api.errorTip'), content: timeoutMsg });
-    } else if (options.errorMessageMode === 'message') {
-      createMessage.error(timeoutMsg);
-    }
+    // if (options.errorMessageMode === 'modal') {
+    //   createErrorModal({ title: t('sys.api.errorTip'), content: timeoutMsg });
+    // } else if (options.errorMessageMode === 'message') {
+    //   createMessage.error(timeoutMsg);
+    // }
 
-    throw new Error(timeoutMsg || t('sys.api.apiRequestFailed'));
+    // throw new Error(timeoutMsg || t('sys.api.apiRequestFailed'));
   },
 
   // 请求之前处理config
@@ -242,9 +253,18 @@ function createAxios(opt?: Partial<CreateAxiosOptions>) {
 export const defHttp = createAxios();
 
 // other api url
-// export const otherHttp = createAxios({
-//   requestOptions: {
-//     apiUrl: 'xxx',
-//     urlPrefix: 'xxx',
-//   },
-// });
+export const otherHttp = createAxios({
+  headers: { 'Content-Type': ContentTypeEnum.JSON },
+  timeout: 10 * 1000,
+  transform,
+  requestOptions: {
+    apiUrl: '/front-api',
+    urlPrefix: '/front',
+    isTransformResponse: true,
+    isReturnNativeResponse: false,
+    withToken: false,
+    ignoreCancelToken: true,
+    joinTime: true,
+    formatDate: true,
+  },
+});