+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# 0.0.0 (2021-10-31)
+
+
+### Features
+
+* Add eslint and prettier configuration ([6b9c11a](https://github.com/lhj-web/edu-mis/commit/6b9c11a4c106f53bad475ae0eaddf3a0b0406b73))
+* Add stylelint configuration ([c9c8022](https://github.com/lhj-web/edu-mis/commit/c9c80229171d816a5fbf66b6bb1d520c8c405ba9))
+* Add SvgIcon and Icon component ([8628357](https://github.com/lhj-web/edu-mis/commit/8628357ba33fd014c468144eb3e63b3711e9e93f))
+* configure vite.config.ts ([1dfa98d](https://github.com/lhj-web/edu-mis/commit/1dfa98d5db1d8d6b5a6f5fde0048b1095b4566e0))
+* Encapsulate axios ([8d3c22c](https://github.com/lhj-web/edu-mis/commit/8d3c22cb36b0c0d41df451fdbc96d8382507336e))
+* Encapsulate axios and configure i18n and add local cache ([f9af257](https://github.com/lhj-web/edu-mis/commit/f9af257ca2f985cd435ceebe359da7dd094e2048))
+* import the windicss and ant-design-vue ([b7b9150](https://github.com/lhj-web/edu-mis/commit/b7b915049b6de1c899dd6129615c57bfcae3e9eb))
+* Initialize the project ([e26e319](https://github.com/lhj-web/edu-mis/commit/e26e319d1c6ea82111b4646f00446ef04921ee7e))
+
+
+
 # 0.0.0 (2021-10-23)